Break disjoint_embeddings into three different modules
