+++ minorminer/utils/parallel_embeddings.py
@@ -25,173 +25,8 @@ from typing import Union
 from collections import Counter
 
 from minorminer.subgraph import find_subgraph
-
-
-def visualize_embeddings(
-    G: nx.Graph,
-    embeddings: list,
-    S: nx.Graph = None,
-    seed: Union[int, np.random.RandomState, np.random.Generator] = None,
-    one_to_iterable: bool = False,
-    **kwargs,
-) -> None:
-    """Visualizes the embeddings using dwave_networkx's layout functions.
-
-    This function visualizes embeddings of a source graph onto a target graph
-    using specialized layouts for structured graphs (chimera, pegasus, or zephyr)
-    or general layouts for unstructured graphs. Node and edge colors are used
-    to differentiate embeddings.
-
-    Args:
-        G: The target graph to be visualized. If the graph
-            represents a specialized topology, it must be constructed using
-            dwave_networkx (e.g., chimera, pegasus, or zephyr graphs).
-        embeddings: A list of embeddings where each embedding is a dictionary
-            mapping nodes of the source graph to nodes in the target graph.
-        S: The source graph to visualize (optional). If provided, only edges
-            corresponding to the source graph embeddings are visualized.
-        seed: A seed for the pseudo-random number generator. When provided,
-            it randomizes the colormap assignment for embeddings.
-        one_to_iterable: Specifies how embeddings are interpreted. Set to `True`
-            to allow multiple target nodes to map to a single source node.
-            Defaults to `False` for one-to-one embeddings.
-        **kwargs: Additional keyword arguments passed to the drawing functions
-            (e.g., `node_size`, `font_size`, `width`).
-
-    Draws:
-        - Specialized layouts: Uses dwave_networkx's `draw_chimera`,
-          `draw_pegasus`, or `draw_zephyr` functions if the graph family is identified.
-        - General layouts: Falls back to networkx's `draw_networkx` for
-          graphs with unknown topology.
-    """
-    fig = plt.gcf()
-    ax = plt.gca()
-    cmap = plt.get_cmap("turbo").copy()
-    cmap.set_bad("lightgrey")
-
-    # Create node color mapping
-    node_color_dict = {q: float("nan") for q in G.nodes()}
-
-    if seed is not None:
-        _embeddings = embeddings.copy()
-        prng = np.random.default_rng(seed)
-        prng.shuffle(_embeddings)
-    else:
-        _embeddings = embeddings
-
-    if one_to_iterable:
-        node_color_dict.update(
-            {
-                q: idx
-                for idx, emb in enumerate(_embeddings)
-                for c in emb.values()
-                for q in c
-            }
-        )
-    else:
-        node_color_dict.update(
-            {q: idx for idx, emb in enumerate(_embeddings) for q in emb.values()}
-        )
-
-    # Create edge color mapping
-    edge_color_dict = {}
-    highlight_edges = []
-    if S is not None:
-        for idx, emb in enumerate(_embeddings):
-            for u, v in S.edges():
-                if u in emb and v in emb:
-                    if one_to_iterable:
-                        targets_u = emb[u]
-                        targets_v = emb[v]
-                    else:
-                        targets_u = [emb[u]]
-                        targets_v = [emb[v]]
-                    for tu in targets_u:
-                        for tv in targets_v:
-                            if G.has_edge(tu, tv):
-                                edge_color_dict[(tu, tv)] = idx
-                                highlight_edges.append((tu, tv))
-    else:
-        for v1, v2 in G.edges():
-            if node_color_dict[v1] == node_color_dict[v2]:
-                edge_color_dict[(v1, v2)] = node_color_dict[v1]
-            else:
-                edge_color_dict[(v1, v2)] = float("nan")
-
-    # Default drawing arguments
-    draw_kwargs = {
-        "G": G,
-        "node_color": [node_color_dict[q] for q in G.nodes()],
-        "edge_color": "lightgrey",
-        "node_shape": "o",
-        "ax": ax,
-        "with_labels": False,
-        "width": 1,
-        "cmap": cmap,
-        "edge_cmap": cmap,
-        "node_size": 300 / np.sqrt(G.number_of_nodes()),
-    }
-    draw_kwargs.update(kwargs)
-
-    topology = G.graph.get("family")
-    # Draw the combined graph with color mappings
-    if topology == "chimera":
-        pos = dnx.chimera_layout(G)
-        dnx.draw_chimera(**draw_kwargs)
-    elif topology == "pegasus":
-        pos = dnx.pegasus_layout(G)
-        dnx.draw_pegasus(**draw_kwargs)
-    elif topology == "zephyr":
-        pos = dnx.zephyr_layout(G)
-        dnx.draw_zephyr(**draw_kwargs)
-    else:
-        pos = nx.spring_layout(G)
-        nx.draw_networkx(**draw_kwargs)
-
-    # Recolor specific edges on top of the original graph
-    if S is None:
-        highlight_edges = [e for e in G.edges() if not np.isnan(edge_color_dict[e])]
-    highlight_colors = [edge_color_dict[e] for e in highlight_edges]
-
-    nx.draw_networkx_edges(
-        G,
-        pos=pos,
-        edgelist=highlight_edges,
-        edge_color=highlight_colors,
-        width=1,
-        edge_cmap=cmap,
-        ax=ax,
-    )
-
-
-def shuffle_graph(
-    G: nx.Graph, seed: Union[int, np.random.RandomState, np.random.Generator] = None
-) -> nx.Graph:
-    """Shuffle the node and edge ordering of a networkx graph.
-
-    For embedding methods that operate as a function of the node or edge
-    ordering this can force diversification in the returned embeddings. Note
-    that special orderings that encode graph structure or geometry may enhance
-    embedder performance (shuffling may lead to longer search times).
-
-    Args:
-        G: A networkx graph
-        seed: When provided, is used to shuffle the order of nodes and edges in
-        the source and target graph. This can allow sampling from otherwise deterministic routines.
-    Returns:
-        nx.Graph: The same graph with modified node and edge ordering.
-    """
-    prng = np.random.default_rng(seed)
-    nodes = list(G.nodes())
-    prng.shuffle(nodes)
-    edges = list(G.edges())
-    prng.shuffle(edges)
-    _G = nx.Graph()
-    _G.add_nodes_from(nodes)
-    _G.add_edges_from(edges)
-    return _G
-
-
+from embeddings import (shuffle_graph, embeddings_to_ndarray, visualize_embeddings)
+from feasibility import (embedding_feasibility_filter, lattice_size_lower_bound, lattice_size_upper_bound)
 def find_multiple_embeddings(
     S: nx.Graph,
     T: nx.Graph,
@@ -277,203 +112,6 @@ def find_multiple_embeddings(
                 _T.remove_nodes_from(emb.values())
         embs.append(emb)
     return embs
-
-
-def embedding_feasibility_filter(
-    S: nx.Graph, T: nx.Graph, one_to_one: bool = False
-) -> bool:
-    """Feasibility filter for embedding.
-
-    Determines if the source graph `S` can be subgraph-embedded onto the target
-    graph `T` based on their degree distributions. Returns `False` if embedding
-    is definitely infeasible; otherwise, returns `True`. False positives are
-    permissible because deciding the graph isomorphism problem is NP-complete,
-    and this filter is designed to be efficient.
-
-    The degree distribution test is a heuristic; stronger filters are possible
-    by exploiting additional graph structure.
-
-    Args:
-        S: The source graph to embed.
-        T: The target graph in which to embed.
-        one_to_one: If True, only 1-to-1 (subgraph) embeddings are allowed.
-            Defaults to False, permitting minor embeddings.
-
-    Returns:
-        bool: `False` if subgraph embedding is definitely infeasible, `True`
-            otherwise.
-    """
-    if T == S:
-        return True
-    if (
-        T.number_of_nodes() < S.number_of_nodes()
-        or T.number_of_edges() < S.number_of_edges()
-    ):
-        return False
-    else:
-        S_degree = np.sort([S.degree[n] for n in S.nodes()])
-        T_degree = np.sort([T.degree[n] for n in T.nodes()])
-
-        if np.any(T_degree[-len(S_degree) :] < S_degree):
-            if one_to_one or T_degree[-1] <= 2:
-                # Too many high degree nodes in S
-                return False
-            else:  # Attempt minor embed (enhance T degrees)
-                # Minor embedding feasibility reduces to bin packing when
-                # considering a best target case knowing only the degree
-                # distribution. In general feasibility is NP-complete, a cheap
-                # marginal degree distribution filter is used.
-
-                # We can eliminate nodes of equal degree assuming best case:
-                ResidualCounts = Counter(T_degree)
-                ResidualCounts.subtract(Counter(S_degree))
-
-                # Target nodes of degree x <=2 are only of use for minor
-                # embedding source nodes of degree <=x:
-                for kS in range(3):
-                    if ResidualCounts[kS] < 0:
-                        ResidualCounts[kS + 1] += ResidualCounts[kS]
-                    ResidualCounts[kS] = 0
-
-                if all(v > 0 for v in ResidualCounts.values()):
-                    return True
-                nT_auxiliary = sum(ResidualCounts.values())
-                if nT_auxiliary < 0:  # extra available to form chains
-                    return False
-
-                # In best case all target nodes have degree kTmax, and chains
-                # are trees. To cover degree k in S requires n auxiliary target
-                # nodes such that kTmax + n(kTmax-2) >= k
-                kTmax = np.max([k for k, v in ResidualCounts.items() if v > 0])
-                min_auxiliary_necessary = sum(
-                    [
-                        -v * np.ceil((k - kTmax) / (kTmax - 2))
-                        for k, v in ResidualCounts.items()
-                        if v < 0
-                    ]
-                )
-                return min_auxiliary_necessary <= nT_auxiliary
-        else:
-            return True
-
-
-def lattice_size_upper_bound(T: nx.Graph = None) -> int:
-    """Determines an upper bound on the lattice size.
-
-    The lattice size is the parameter ``m`` of a dwave_networkx graph, also
-    called number of rows.
-
-    Args:
-        T: The target graph in which to embed. The graph must be of type
-            zephyr, pegasus or chimera and constructed by dwave_networkx.
-    Returns:
-        int: The maximum possible size of a tile
-    """
-    return max(T.graph.get("rows"), T.graph.get("columns"))
-
-
-def lattice_size_lower_bound(
-    S: nx.Graph,
-    T: nx.Graph = None,
-    topology: str = None,
-    t: int = None,
-    one_to_one: bool = False,
-) -> float:
-    """Returns a lower bound on the size necessary for embedding.
-
-    The lattice size is the parameter `m` of a dwave_networkx graph, also
-    called number of rows. The function returns a lower bound (necessary but
-    not sufficient for embedding) using efficiently established graph
-    properties such as the number of nodes, number of edges, node-degree
-    distribution, and two-colorability
-
-    Args:
-        S: The source graph to embed.
-        T: The target graph in which to embed. The graph must be of type
-            'zephyr', 'pegasus', or 'chimera' and constructed by dwave_networkx.
-        topology: The topology ('chimera', 'pegasus', or 'zephyr'). This is
-            inferred from `T` by default. Any set value must be consistent with
-            `T` (if `T` is not None).
-        t: The tile parameter, relevant for 'zephyr' and 'chimera' topologies.
-            Inferred from `T` by default. Any set value must be consistent with
-            `T` (if `T` is not None).
-        one_to_one: True if a subgraph embedding is assumed, False for general
-            embeddings.
-
-    Raises:
-        ValueError: If `T` and `topology` are inconsistent, or if `T` and `t` are inconsistent.
-        ValueError: If neither `T` nor `topology` is specified.
-        ValueError: If `S` cannot be embedded in `T` or the specified topology.
-
-    Returns:
-        float: Minimum `lattice_size` for embedding to be feasible. Returns
-            None if embedding for any number of graph rows is infeasible.
-    """
-    # Comment, similar bounds are possible allowing for minor embeddings,
-    # the degree distribution is bounded as chain length increases, whilst
-    # the number of nodes is depleted.
-    # This could be a possible feature expansion.
-    if T is not None:
-        if embedding_feasibility_filter(S, T, one_to_one) is False:
-            return None
-        if topology is None:
-            topology = T.graph.get("family")
-        elif topology != T.graph.get("family"):
-            raise ValueError("Arguments T and topology are inconsistent")
-        if t is None:
-            t = T.graph["tile"]
-    else:
-        if topology is None:
-            raise ValueError("T or topology must be specified")
-        if t is None:
-            t = 4
-        max_degrees = {"chimera": 2 + 2 * t, "pegasus": 15, "zephyr": 4 + 4 * t}
-        max_source_degree = max(S.degree[n] for n in S.nodes())
-        if max_source_degree > max_degrees[topology]:
-            return None
-
-    N = S.number_of_nodes()
-    if topology == "chimera":
-        # Two colorability is necessary, and cheap to check
-        if any(
-            c > 2
-            for c in nx.algorithms.coloring.greedy_coloring.greedy_color(
-                S, strategy="connected_sequential_bfs"
-            ).values()
-        ):
-            return None
-
-        def generator(lattice_size):
-            return dnx.chimera_graph(m=lattice_size, n=lattice_size, t=t)
-
-        # A lower bound based on number of variables N = m*n*2*t
-        lattice_size = np.ceil(np.sqrt(N / 4 / t))
-    elif topology == "pegasus":
-
-        def generator(lattice_size):
-            return dnx.pegasus_graph(m=lattice_size)
-
-        # A lower bound based on number of variables N = (m*24-8)*(m-1)
-        lattice_size = np.ceil(1 / 12 * (8 + np.sqrt(6 * N + 16)))
-    elif topology == "zephyr":
-
-        def generator(lattice_size):
-            return dnx.zephyr_graph(m=lattice_size, t=t)
-
-        # A lower bound based on number of variables N = (2m+1)*m*4*t
-        lattice_size = np.ceil((np.sqrt(2 * N / t + 1) - 1) / 4)
-    else:
-        raise ValueError(
-            "source graphs must be a graph constructed by "
-            "dwave_networkx as chimera, pegasus or zephyr type"
-        )
-    # Evaluate tile feasibility (defect free subgraphs)
-    lattice_size = round(lattice_size)
-    tile = generator(lattice_size=lattice_size)
-    while embedding_feasibility_filter(S, tile, one_to_one) is False:
-        lattice_size += 1
-        tile = generator(lattice_size=lattice_size)
-    return lattice_size
 
 
 def find_sublattice_embeddings(
@@ -644,34 +282,6 @@ def find_sublattice_embeddings(
     return embs
 
 
-def embeddings_to_ndarray(embs: list, node_order=None):
-    """Convert list of embeddings into an ndarray
-
-    Note this assumes the target graph is labeled by integers and the embedding
-    is 1 to 1(numeric) in all cases. This is the format returned by
-    minorminor.subgraph for the standard presentation of QPU graphs.
-
-    Args:
-        embs: A list of embeddings, each list entry in the
-            form of a dictionary with integer values.
-        node_order: An iterable giving the ordering of
-            variables in each row. When not provided variables are ordered to
-            match the first embedding :code:`embs[0].keys()`
-
-    Raises:
-        ValueError: If `embs` is empty and `node_order` cannot be inferred.
-
-    Returns:
-        np.ndarray: An embedding matrix; each row defines an embedding ordered
-            by node_order.
-    """
-    if node_order is None:
-        if len(embs) == 0:
-            raise ValueError("shape of ndarray cannot be inferred")
-        else:
-            node_order = embs[0].keys()
-
-    return np.asarray([[ie[v] for ie in embs] for v in node_order]).T
 
 
 if __name__ == "__main__":
